01 28 2025
Had a great time exploring an answer.

I learned that I was placing the wrong emphasis on the MAX clause.

I haven't submitted it yet.  I will do that next time.

For now, I believe that I didn't pick up on the fact that the LAT_N had to be the MAX where as I was using the MAX on LONG_W.

Such weird wording on these questions sometimes.